# api/openapi/swagger.yaml
openapi: 3.0.0
info:
  title: backend service
  description: Сервис для управления ПВЗ и приемкой товаров
  version: 1.0.0

# Добавляем секцию servers для удобства тестирования в Swagger UI/Postman
servers:
  - url: http://localhost:8080 # URL вашего запущенного сервиса
    description: Локальный сервер разработки
components:
  schemas:
...
    PVZ:
      description: Пункт выдачи заказов
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор ПВЗ
          readOnly: true # ID генерируется сервером
        registrationDate:
          type: string
          format: date-time
          description: Дата и время регистрации ПВЗ
          readOnly: true # Устанавливается сервером
        city:
          $ref: '#/components/schemas/PVZCity' # Ссылка на Enum
      required: [city] # Только город обязателен при создании
    PVZCity: # Выносим Enum в отдельную схему
      type: string
      description: Город расположения ПВЗ
      enum: [Москва, Санкт-Петербург, Казань]
      example: Москва
    Reception:
      description: Запись о приемке товаров
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор приемки
          readOnly: true
        dateTime:
          type: string
          format: date-time
          description: Дата и время начала приемки
          readOnly: true
        pvzId:
          type: string
          format: uuid
          description: ID пункта выдачи заказов, к которому относится приемка
        status:
          $ref: '#/components/schemas/ReceptionStatus' # Ссылка на Enum
      required: [dateTime, pvzId, status] # Все поля, кроме ID, обязательны при ответе
    ReceptionStatus: # Выносим Enum в отдельную схему
      type: string
      description: Статус приемки товаров
      enum: [in_progress, closed] # Используем 'closed' вместо 'close' для единообразия с кодом
      example: in_progress
    Product:
      description: Товар, принятый в ПВЗ
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор товара
          readOnly: true
        dateTimeAdded: # Меняем dateTime на dateTimeAdded для соответствия домену
          type: string
          format: date-time
          description: Дата и время добавления товара в приемку
          readOnly: true
        type:
          $ref: '#/components/schemas/ProductType' # Ссылка на Enum
        receptionId:
          type: string
          format: uuid
          description: ID приемки, к которой относится товар
      required: [type, receptionId] # Обязательны при создании/ответе
    ProductType: # Выносим Enum в отдельную схему
      type: string
      description: Тип товара
      enum: [электроника, одежда, обувь]
      example: одежда
    Error:
      description: Стандартный ответ с ошибкой
      type: object
      properties:
        message:
          type: string
          description: Текстовое описание ошибки
      required: [message]

    # --- Схемы для составных ответов ---
    ProductInfo: # Информация о товаре для списка (если нужна отдельная)
      description: Краткая информация о товаре в рамках ответа списка ПВЗ
      allOf: # Наследуем все от Product
        - $ref: '#/components/schemas/Product'
      # Можно добавить/убрать поля при необходимости
    ReceptionInfo: # Информация о приемке с товарами для списка
      description: Информация о приемке, включая список товаров, для ответа списка ПВЗ
      type: object
      properties:
        reception:
          $ref: '#/components/schemas/Reception'
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductInfo' # Ссылка на схему товара для списка
      required:
        - reception
        - products
    PvzListItem: # Элемент списка ПВЗ
      description: Один элемент в списке ПВЗ, включая ПВЗ и его приемки
      type: object
      properties:
        pvz:
          $ref: '#/components/schemas/PVZ'
        receptions:
          type: array
          items:
            $ref: '#/components/schemas/ReceptionInfo' # Ссылка на приемку с товарами
      required:
        - pvz
        - receptions
    PvzListResponse: # Полный ответ для GET /pvz
       description: Ответ со списком ПВЗ, пагинацией и деталями приемок/товаров
       type: object
       properties:
         items:
           type: array
           items:
             $ref: '#/components/schemas/PvzListItem' # Ссылка на элемент списка
         totalCount:
           type: integer
           description: Общее количество ПВЗ, доступных по запросу (без учета пагинации)
           example: 42
         page:
           type: integer
           description: Номер текущей страницы (начиная с 1)
           example: 1
         limit:
           type: integer
           description: Количество элементов на странице
           example: 10
       required:
         - items
         - totalCount
         - page
         - limit

    # --- Схемы для тел запросов (где они отличаются от основных схем) ---
...
  securitySchemes:
    bearerAuth: # Имя схемы должно совпадать с используемым в security: - bearerAuth: []
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен доступа, полученный через /login или /dummyLogin
paths:
...
  /pvz:
    post:
      summary: Создание ПВЗ (только для модераторов)
      operationId: postPvz
      tags:
        - PVZ
      security:
        - bearerAuth: [] # Указываем схему безопасности
      requestBody:
        description: Данные для создания нового ПВЗ (только город)
        required: true
        content:
          application/json:
            schema:
              # Ссылка на основную схему PVZ, но при создании используется только city
              # Можно создать отдельную CreatePVZRequest схему, если нужно
              $ref: '#/components/schemas/PVZ'
      responses:
        '201':
          description: ПВЗ успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PVZ' # Возвращаем созданный ПВЗ
        '400':
          description: Неверный запрос (например, недопустимый город)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': # Добавляем 401, т.к. эндпоинт защищен
          description: Неавторизован (токен отсутствует, невалиден или истек)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен (недостаточно прав, не модератор)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Получение списка ПВЗ с деталями
      description: Получение списка ПВЗ с возможностью фильтрации по дате приемки и пагинацией. Включает информацию о приемках и товарах.
      operationId: getPvzList
      tags:
        - PVZ
      security:
        - bearerAuth: [] # Защищаем эндпоинт
      parameters:
        - name: startDate
          in: query
          description: Начальная дата диапазона (фильтр для приемок)
          required: false
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Конечная дата диапазона (фильтр для приемок)
          required: false
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          description: Номер страницы
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 30 # Оставляем ограничение
            default: 10
      responses:
        '200':
          description: Успешный ответ со списком ПВЗ и деталями
          content:
            application/json:
              schema:
                # Ссылка на схему полного ответа
                $ref: '#/components/schemas/PvzListResponse'
        '401': # Добавляем 401
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400': # Добавляем 400 для невалидных параметров пагинации/дат
          description: Неверный запрос (например, невалидный формат даты или номер страницы)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...